Fix monorail selection
