+++ src/main/java/com/railwayteam/railways/content/coupling/TrainUtils.java
@@ -114,6 +114,9 @@ public class TrainUtils {
      * Adds the carriages of backTrain onto the end of frontTrain.
      */
     public static Train combineTrains(Train frontTrain, Train backTrain, Vec3 itemDropPos, Level itemDropLevel, int carriageSpacing) {
+        if (frontTrain.derailed || backTrain.derailed) {
+            return frontTrain;
+        }
         int frontTrainSize = frontTrain.carriages.size();
         frontTrain.carriages.addAll(backTrain.carriages);
         backTrain.carriages.clear();
